feat: 提取公共 UserSearchSelect 组件，统一用户搜索体验
- 新增 UserSearchSelect 公共组件：头像 + 昵称 + @username + 角色徽章 + 活跃时间
- 组件支持 sm/md 尺寸、自动获取用户列表、可搜索过滤
- SystemLogsTab: 替换用户下拉为 UserSearchSelect，应用下拉加 icon + 可搜索，
  状态下拉加 icon + 中文 + emoji，所有下拉加 leftIcon，statusBadge 显示状态码 + 中文 + 彩色背景
- LlmLogsPage: 替换用户 SearchableSelect 为 UserSearchSelect
- DataTransferPage: 删除 200 行本地 SearchableUserPicker，改用公共组件
- 注册组件到 prd-admin/CLAUDE.md
